Fix to #6383 - The wrong type is being selected when an extension method is used in a subquery
Problem was that when translating collection navigation outside projection and AdditionalFromClause we assumed it's always going to be wrapped in a subquery.
This is the case for the most part because in order to be useful inside a query, collection needs to have it's cardinality reduced to single element using a result operator (Count, Any, FirstOrDefault etc), which forces subquery on the collection.
However it is also possible to use client method, which takes collection and returns single element. For this case, subquery won't be generated and we were unable to translate such query.

Fix is to inject subqueries where needed, using same mechanism we use in case of projection. Client method forces materialization on the collection, so it's safe to use.