Fix to #8060 - Query: Make our reflection info selection logic deterministic
Problem was that the way we extract MethodInfos for translators was always deterministic and could lead to errors if the new overloads of those methods were to be added.

Fix is to use pattern in which we fully specify the name and types of all arguments to the method - this is guaranteed to be unique.